feat: :sparkles: Controller update, data with images.

+++ src/controllers/categories.controller.ts
@@ -1,18 +1,13 @@
-import { Request, Response } from "express";
+import { Request, Response } from "express"
 import { CategorieDTO } from "../DTOs/categories.dto"
-import * as service from "../services/categories.service";
-
+import * as service from "../services/categories.service"
 
 // CREATE
 export const create = async (req: Request, res: Response) => {
     try {
         const categorieData = CategorieDTO.create(req.body)
         const categorie = await service.createCategories(categorieData)
-
-        res.status(201).json({
-            message: "Categorie created correctly",
-            categorie
-        })
+        res.status(201).json({message: "Categorie created correctly",categorie})
     } catch (error: any) {
         res.status(400).json({message: error.message || "Invalid data"})
     }
@@ -21,25 +16,27 @@ export const create = async (req: Request, res: Response) => {
 // GET ALL
 export const getAll = async (_req: Request, res: Response) => {
     try {
-        const categories = await service.getCategories();
-        res.json(categories);
+        const categories = await service.getCategories()
+        res.status(200).json(categories)
     } catch (error) {
-        res.status(500).json({ message: "Error fetching categories", error });
+        res.status(500).json({message: "Error fetching categories"})
     }
 }
 
 // GET BY ID
 export const getById = async (req: Request, res: Response) => {
     try {
-        const id = Number(req.params.id);
-        const categories = await service.getCategoriesById(id);
-        if (!categories) {
-            res.status(404).json({ message: "Categorie not found" });
-            return;
+        const id = Number(req.params.id)
+        if (isNaN(id)) {
+            return res.status(400).json({ message: "Invalid ID" })
         }
-        res.json(categories);
+        const categorie = await service.getCategoriesById(id)
+        if (!categorie) {
+            return res.status(404).json({ message: "Categorie not found" })
+        }
+        res.status(200).json(categorie)
     } catch (error) {
-        res.status(500).json({ message: "Error fetching categories", error });
+        res.status(500).json({message: "Error fetching categorie"})
     }
 }
 
@@ -48,13 +45,12 @@ export const update = async (req: Request, res: Response) => {
     try {
         const id = Number(req.params.id)
         if (isNaN(id)) {
-            return res.status(400).json({message: "Invalid ID"})
+            return res.status(400).json({ message: "Invalid ID" })
         }
-        const CategorieData = CategorieDTO.update(req.body)
-        const categorie = await service.updateCategories(id, CategorieData)
-
+        const categorieData = CategorieDTO.update(req.body)
+        const categorie = await service.updateCategories(id, categorieData)
         if (!categorie) {
-            return res.status(404).json({message: "Categorie not found"})
+            return res.status(404).json({ message: "Categorie not found" })
         }
         res.status(200).json({message: "Categorie updated correctly",categorie})
     } catch (error: any) {
@@ -65,15 +61,16 @@ export const update = async (req: Request, res: Response) => {
 // DELETE
 export const remove = async (req: Request, res: Response) => {
     try {
-        const id = Number(req.params.id);
-        const deletedCategories = await service.deleteCategories(id);
-
-        if (!deletedCategories) {
-            res.status(404).json({ message: "Categories not found" });
-            return;
+        const id = Number(req.params.id)
+        if (isNaN(id)) {
+            return res.status(400).json({ message: "Invalid ID" })
         }
-        res.json(deletedCategories);
+        const deletedCategorie = await service.deleteCategories(id)
+        if (!deletedCategorie) {
+            return res.status(404).json({ message: "Categorie not found" })
+        }
+        res.status(200).json({message: "Categorie deleted correctly",categorie: deletedCategorie})
     } catch (error) {
-        res.status(500).json({ message: "Error deleting categories", error });
+        res.status(500).json({message: "Error deleting categorie"})
     }
 }
